backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
scikit-learn
numpy
pypdf